+++ commands/base/custom_command.py
@@ -1,35 +1,5 @@
 import abc
-import os
-
-from qgis.core import QgsVectorLayer, QgsMapLayerRegistry, QgsVectorJoinInfo
-
 from ThreeDiToolbox.utils.user_messages import pop_up_info
-
-
-def join_stats(filepath, layer, view_layer_field, csv_field='id'):
-    """Join the generated stats csv with the layer
-
-    Args:
-        filepath: path to the csv file
-        layer: the layer we want to join the csv with
-        view_layer_field: the id (e.g. primary key) of layer
-        csv_field: the id of the csv layer (which is always 'id' in the case
-            of the CustomCommand scripts)
-    """
-    filename = os.path.basename(filepath)
-    csv_layer_name = os.path.splitext(filename)[0]
-    csv_uri = "file:///" + filepath
-    csv_layer = QgsVectorLayer(csv_uri, csv_layer_name, "delimitedtext")
-    QgsMapLayerRegistry.instance().addMapLayer(csv_layer)
-    join_info = QgsVectorJoinInfo()
-    join_info.joinLayerId = csv_layer.id()
-    join_info.joinFieldName = csv_field
-    join_info.targetFieldName = view_layer_field
-    join_info.memoryCache = True
-    layer.addJoin(join_info)
-
-    pop_up_info("Finished joining '%s' with '%s'." % (
-        csv_layer_name, layer.name()), title='Join finished')
 
 
 class CustomCommandBase(object):