+++ qtdebug.py
@@ -1,6 +1,7 @@
 """Qt/QGIS debugging tools"""
 
-def pyqt_set_trace():
+
+def set_trace():
     """Set a breakpoint in the Python debugger that works with Qt/QGIS.
 
     You need to start QGIS from the terminal for this.